requirements.txt
# Python requirements for the login application
# Install with: pip install -r requirements.txt

# MySQL database connector
mysql-connector-python>=8.0.0

# PyQt6 for GUI (if not already installed)
PyQt6>=6.0.0

# Additional dependencies that might be needed
# (These are usually included with Python standard library)
# json - built-in
# os - built-in
# time - built-in
# hashlib - built-in
# secrets - built-in
# logging - built-in
# getpass - built-in
